requirements.txt
openai-whisper
pytube
pydub
openai
pdf2docx